cleanup command bus
